+++ src/interfaces.ts
@@ -34,13 +34,13 @@ import ISQExpr = powerbi.data.ISQExpr;
 import { interactivitySelectionService as interactivityService } from "powerbi-visuals-utils-interactivityutils";
 import SelectableDataPoint = interactivityService.SelectableDataPoint;
 
-import { ChicletSlicerSettings } from "./settings";
+import { ChicletSlicerSettingsModel } from "./chicletSlicerSettingsModel";
 
 export interface ChicletSlicerData {
     categorySourceName: string;
     formatString: string;
     slicerDataPoints: ChicletSlicerDataPoint[];
-    slicerSettings: ChicletSlicerSettings;
+    slicerSettings: ChicletSlicerSettingsModel;
     hasSelectionOverride?: boolean;
     hasHighlights: boolean;
     identityFields: ISQExpr[];